# [11.x] Optimize container by caching reflection data