--- a/src/Illuminate/Container/ClassDefinition/ClassDefinitionProvider.php
+++ b/src/Illuminate/Container/ClassDefinition/ClassDefinitionProvider.php
@@ -1,0 +1,47 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Illuminate\Container\ClassDefinition;
+
+use Illuminate\Container\Util;
+use ReflectionClass;
+use ReflectionParameter;
+
+final class ClassDefinitionProvider
+{
+    /**
+     * @var array<string, ClassDefinition>
+     */
+    private array $definitions = [];
+
+    public function get(string $className): ClassDefinition
+    {
+        if (isset($this->definitions[$className])) {
+            return $this->definitions[$className];
+        }
+
+        $reflection = new ReflectionClass($className);
+        $definition = new ClassDefinition(
+            class: $className,
+            isInstantiable: $reflection->isInstantiable(),
+            isConstructorDefined: !is_null($reflection->getConstructor()),
+            parameters: array_map(
+                fn (ReflectionParameter $parameter) => new Parameter(
+                    name: $parameter->getName(),
+                    className: Util::getParameterClassName($parameter),
+                    isVariadic: $parameter->isVariadic(),
+                    isDefaultValueAvailable: $parameter->isDefaultValueAvailable(),
+                    defaultValue: $parameter->isDefaultValueAvailable() ? $parameter->getDefaultValue() : null,
+                    declaringClassName: $parameter->getDeclaringClass()->getName(),
+                    asString: (string) $parameter,
+                ),
+                $reflection->getConstructor()?->getParameters() ?? [],
+            ),
+        );
+
+        $this->definitions[$className] = $definition;
+
+        return $definition;
+    }
+}