--- a/tests/Database/DatabaseMigrationRefreshCommandTest.php
+++ b/tests/Database/DatabaseMigrationRefreshCommandTest.php
@@ -121,6 +121,8 @@ class ApplicationDatabaseRefreshStub extends Application
 {
     public function __construct(array $data = [])
     {
+        parent::__construct();
+
         foreach ($data as $abstract => $instance) {
             $this->instance($abstract, $instance);
         }