--- a/tests/Database/DatabaseMigrationResetCommandTest.php
+++ b/tests/Database/DatabaseMigrationResetCommandTest.php
@@ -80,6 +80,8 @@ class ApplicationDatabaseResetStub extends Application
 {
     public function __construct(array $data = [])
     {
+        parent::__construct();
+
         foreach ($data as $abstract => $instance) {
             $this->instance($abstract, $instance);
         }