It's another slight improvement for large applications with many dependencies. In Laravel we can define services with interfaces as scoped or singleton to force single creation. But for services without an interface (or those declared by using the bind method), there is no need to use scoped or singleton, so every time the service is needed a new instance needs to be created, and in the case of complex applications it could be quite slow. As the first improvement, we can cache data retrieved by using reflection to make those new instances a little bit faster. For a small number of services, it isn't a big optimization, but for a larger number of services, it could mean a few milliseconds per request, which could be a nice optimization. 

**Benchmarks**

It was tested on the example application with a few nested dependencies similar to the following structure: 
.Service1
..Service11
....ServiceA
....ServiceB
....ServiceC
....ServiceD
..Service12
....ServiceA
....ServiceB
....ServiceC
....ServiceD
..Service13
....ServiceA
....ServiceB
....ServiceC
....ServiceD
..Service14
....ServiceA
....ServiceB
....ServiceC
....ServiceD
..Service15
....ServiceA
....ServiceB
....ServiceC
....ServiceD

**BEFORE**
```
ab -n 1000 -c 1 http://127.0.0.1:8080/
This is ApacheBench, Version 2.3 <$Revision: 1903618 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient)
Completed 100 requests
Completed 200 requests
Completed 300 requests
Completed 400 requests
Completed 500 requests
Completed 600 requests
Completed 700 requests
Completed 800 requests
Completed 900 requests
Completed 1000 requests
Finished 1000 requests


Server Software:        nginx/1.25.5
Server Hostname:        127.0.0.1
Server Port:            8080

Document Path:          /
Document Length:        2 bytes

Concurrency Level:      1
Time taken for tests:   44.335 seconds
Complete requests:      1000
Failed requests:        0
Total transferred:      1128000 bytes
HTML transferred:       2000 bytes
Requests per second:    22.56 [#/sec] (mean)
Time per request:       44.335 [ms] (mean)
Time per request:       44.335 [ms] (mean, across all concurrent requests)
Transfer rate:          24.85 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.3      0       8
Processing:    34   44   9.8     41     133
Waiting:       34   44   9.8     41     133
Total:         35   44   9.8     41     133

Percentage of the requests served within a certain time (ms)
  50%     41
  66%     43
  75%     45
  80%     47
  90%     55
  95%     65
  98%     73
  99%     81
 100%    133 (longest request)
```
<img width="1916" alt="Screenshot 2024-06-13 at 15 40 06" src="https://github.com/laravel/framework/assets/6762480/47dabea5-2c2f-4a4f-8a8c-49dfecee9104">


**AFTER**

```
ab -n 1000 -c 1 http://127.0.0.1:8080/
This is ApacheBench, Version 2.3 <$Revision: 1903618 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient)
Completed 100 requests
Completed 200 requests
Completed 300 requests
Completed 400 requests
Completed 500 requests
Completed 600 requests
Completed 700 requests
Completed 800 requests
Completed 900 requests
Completed 1000 requests
Finished 1000 requests


Server Software:        nginx/1.25.5
Server Hostname:        127.0.0.1
Server Port:            8080

Document Path:          /
Document Length:        2 bytes

Concurrency Level:      1
Time taken for tests:   35.608 seconds
Complete requests:      1000
Failed requests:        0
Total transferred:      1128000 bytes
HTML transferred:       2000 bytes
Requests per second:    28.08 [#/sec] (mean)
Time per request:       35.608 [ms] (mean)
Time per request:       35.608 [ms] (mean, across all concurrent requests)
Transfer rate:          30.94 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.2      0       4
Processing:    29   35   7.9     33     110
Waiting:       29   35   7.9     33     110
Total:         29   35   7.9     33     110

Percentage of the requests served within a certain time (ms)
  50%     33
  66%     34
  75%     36
  80%     37
  90%     41
  95%     51
  98%     59
  99%     72
 100%    110 (longest request)
```

<img width="1923" alt="Screenshot 2024-06-13 at 15 40 16" src="https://github.com/laravel/framework/assets/6762480/0307202a-be70-4aa3-b5e2-42d25f7773fd">