@rojtjo 2024-06-13
This looks very promising!
Did you run these benchmarks with OPcache enabled? If not, could you post the results with it enabled?