@sarven 2024-06-14
@rojtjo It was tested with enabled Opcache. Here is the screenshot from Blackfire:
<img width="1964" alt="Screenshot 2024-06-14 at 08 29 43" src="https://github.com/laravel/framework/assets/6762480/7283fad1-4c28-4f8c-87c0-997fa1fea491">