@taylorotwell 2024-06-17
Can we maybe get some real world benchmarks on apps that would simulate a bit more "typical" Laravel application in production with a couple of database queries?