@Tofandel 2024-06-19
I agree, I could give it a try, but right now all my apps are in 10.x, I didn't yet migrate to 11.x, and I'm wondering if the performance might be due to just the change to properties rather than methods in the reflection, because from my previous testing I noticed that php already caches the result of `new ReflectionClass()` because the first call always takes ~20 times more time than subsequent calls