@sarven 2024-06-21
@taylorotwell It doesn't make any sense to create an app with SQL queries, because this change doesn't impact queries, it only impacts the container. I know that it is an imagined example, but I was motivated by seeing something like this in one of the apps that I maintain:
<img width="400" alt="Screenshot 2024-06-21 at 14 57 51" src="https://github.com/laravel/framework/assets/6762480/db695e6d-b52e-437a-a641-91cdeae00dfb">

@Tofandel I've created the repository with a comparison. You can check it here: https://github.com/sarven/laravel-optimization-test

**Before**
```
Time: 19.366332292557ms
Memory: 20.48 KB
```

**After**
```
Time: 15.390226840973ms
Memory: 10.24 KB
```

It's counterintuitive, but seems like in the cached version memory usage is even lower. I don't know how PHP works under the hood, so can't explain that.

I'm thinking if the possibility to flush this cache is required or not. 

BTW I would want to improve the whole container, so this is the first of a few smaller changes. I would want to introduce something like a compilation of the container before deploying the application. It should improve the situation when in large monolithic apps you can see a profile like this:

<img width="388" alt="Screenshot 2024-06-21 at 15 04 45" src="https://github.com/laravel/framework/assets/6762480/4ac90c32-61d1-4527-b550-bb7c7f5b5a98">

where the whole request time is 500ms (Blackfire time, not the real one)