@Tofandel 2026-03-01
@sarven 

This is not how you test memory footprint in php

```php
for ($i = 0; $i < $count; $i++) {
    $app = require __DIR__.'/bootstrap/app.php';
    $app->make(Kernel::class)->bootstrap();

    $start = microtime(true);
    $startMemory = memory_get_usage(true);
    $service = $app->make(RootService::class);
    $service = $app->make(RootService::class);
    $service = $app->make(RootService2::class);
    $service = $app->make(RootService2::class);
    $end = microtime(true);
    $endMemory = memory_get_usage(true);

    $data[] = [
        'time' => ($end - $start) * 1000,
        'memory' => $endMemory - $startMemory,
    ];

}
```

You have no guarantee about when the garbage collector runs, so first you need to run it yourself before the calls to get the memory usage.

You are also testing the memory footprint of a statically cached definition in a loop. This will give you meaningless data because there will be no increase after the first loop. In the real world you would take the memory of one loop and multiply it by the number of loops because php is bootstrapped from scratch every time (unless running on octane)

I would also suggest checking the peak memory as this is what will matter in high request density scenario instead of the memory after garbage collection.